requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
requests==2.32.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
anthropic==0.25.8
beautifulsoup4==4.12.3
lxml==5.2.1
PyPDF2==3.0.1
gunicorn==22.0.0
